Some love for the Kitchen (#6550)
* Some love for the Kitchen

Roach eggs contain eggyolk and blattedin.
Removed the code for making cutlets, replaced with the pizza code instead, allowing chems from cutlets to be inherited.
Biogenerator can now print boxes of eggs.

* Reverted changes to rawmeatball

The rawmeatball is actually a legacy item and is not used, despite having a sprite. I didn't remove it yet in case a later change wants to reuse it.